requirements.txt
accelerate==1.7.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.7
aiohttp-cors==0.8.1
aiosignal==1.3.2
airportsdata==20250523
annotated-types==0.7.0
anthropic==0.52.2
antlr4-python3-runtime==4.9.3
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
astor==0.8.1
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
av==14.4.0
awscli==1.25.97
babel==2.17.0
backports-datetime-fromisoformat==2.0.3
beautifulsoup4==4.13.4
blake3==1.0.5
bleach==6.2.0
blobfile==3.1.0
boto3==1.39.14
botocore==1.39.14
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
codetiming==1.4.0
colorama==0.3.9
coloredlogs==15.0.1
colorful==0.5.6
comm==0.2.2
compressed-tensors==0.9.3
contourpy==1.3.2
cuda-bindings==12.9.0
cuda-python==12.9.0
cupy-cuda12x==13.4.1
cycler==0.12.1
datasets==3.6.0
debugpy==1.8.14
decorator==4.4.2
decord==0.6.0
deepspeed==0.16.4
defusedxml==0.7.1
Deprecated==1.2.18
depyf==0.18.0
dill==0.3.8
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.16
docutils==0.21.2
einops==0.8.1
email_validator==2.2.0
environs==14.2.0
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
fastjsonschema==2.21.2
fastrlock==0.8.3
ffmpy==0.6.0
filelock==3.18.0
fire==0.7.0
fonttools==4.58.4
fqdn==1.5.1
frozenlist==1.6.2
fsspec==2025.3.0
gdown==5.2.0
gguf==0.17.0
gitdb==4.0.12
GitPython==3.1.44
glog==0.3.1
google-api-core==2.25.0
google-auth==2.40.2
googleapis-common-protos==1.70.0
gradio==5.31.0
gradio_client==1.10.1
groovy==0.1.2
grpcio==1.72.1
h11==0.16.0
hf-xet==1.1.3
hf_transfer==0.1.9
hjson==3.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.33.4
humanfriendly==10.0
humanize==4.12.3
hydra-core==1.3.2
identify==2.6.12
idna==3.10
ImageHash==4.3.2
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.0.0
iniconfig==2.1.0
interegular==0.3.3
ipykernel==6.29.5
ipython==8.37.0
ipywidgets==8.1.7
isoduration==20.11.0
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.6
jiter==0.10.0
jmespath==0.10.0
joblib==1.5.1
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.6
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
lark==1.2.2
latex2sympy2_extended==1.10.1
lazy_loader==0.4
librosa==0.11.0
liger_kernel==0.5.10
litellm==1.72.0
llguidance==0.7.26
llvmlite==0.44.0
lm-format-enforcer==0.10.11
lxml==6.0.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==4.0.0
math-verify==0.7.0
mathruler==0.1.0
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
megatron-core @ git+https://github.com/NVIDIA/Megatron-LM.git@408eb7186a68ba4d30ee6cc8b05b4de6ba702148
mistral_common==1.8.1
mistune==3.1.3
modelscope==1.26.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.4.4
multiprocess==0.70.16
multiprocessing-logging==0.3.4
munch==4.0.0
nanobind==2.7.0
narwhals==2.1.2
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
nltk==3.9.1
nodeenv==1.9.1
notebook==7.4.5
notebook_shim==0.2.4
numba==0.61.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
openai==1.84.0
opencensus==0.11.4
opencensus-context==0.1.3
opencv-fixer==0.2.5
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-semantic-conventions-ai==0.4.9
optree==0.16.0
orjson==3.10.18
outlines==0.1.11
outlines_core==0.1.26
overrides==7.7.0
packaging==25.0
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
partial-json-parser==0.2.1.1.post5
peft==0.15.2
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
plotly==6.3.0
pluggy==1.6.0
pooch==1.8.2
pre_commit==4.2.0
proglog==0.1.12
projectaria-tools==1.6.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==4.25.8
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
py-spy==0.4.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.1
pybind11==2.13.6
pycocotools==2.0.7
pycountry==24.6.1
pycparser==2.22
pycryptodomex==3.23.0
pydantic==2.10.6
pydantic-extra-types==2.10.5
pydantic_core==2.27.2
pydub==0.25.1
pyext==0.7
Pygments==2.19.1
pylatexenc==2.10
pynvml==12.0.0
pyparsing==3.2.3
PySocks==1.7.1
pytest==8.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-gflags==3.1.2
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
PyWavelets==1.8.0
PyYAML==6.0.2
pyzmq==26.4.0
qwen-vl-utils==0.0.11
ray==2.46.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rerun-sdk==0.24.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rich-toolkit==0.14.7
rouge-chinese==1.0.3
rpds-py==0.25.1
rsa==3.4.2
ruff==0.11.12
s3transfer==0.13.1
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
semantic-version==2.10.0
Send2Trash==1.8.3
sentencepiece==0.2.0
sentry-sdk==2.29.1
setproctitle==1.3.6
sgl-kernel==0.1.0
sglang==0.4.6.post1
shellingham==1.5.4
shtab==1.7.2
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.7
soxr==0.5.0.post1
sse-starlette==2.3.6
stack-data==0.6.3
starlette==0.46.2
sympy==1.13.1
tensordict==0.6.2
termcolor==3.1.0
terminado==0.18.1
threadpoolctl==3.6.0
tiktoken==0.9.0
tinycss2==1.4.0
tokenizers==0.21.1
tomli==2.2.1
tomlkit==0.13.3
torch==2.6.0
torch_memory_saver==0.0.6
torchao==0.11.0
torchaudio==2.6.0
torchdata==0.11.0
torchvision==0.21.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformer_engine @ git+https://github.com/NVIDIA/TransformerEngine.git@c55e425ac5751c72bfa656560fc7c65ab7d895e9
transformers==4.52.4
triton==3.2.0
trl==0.9.6
typer==0.16.0
types-python-dateutil==2.9.0.20250809
typing-inspection==0.4.1
typing_extensions==4.14.0
tyro==0.8.14
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.34.3
uvloop==0.21.0
virtualenv==20.31.2
vllm==0.8.5
wandb==0.20.0
watchfiles==1.0.5
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
widgetsnbextension==4.0.14
wrapt==1.17.2
xformers==0.0.29.post2
xgrammar==0.1.18
xxhash==3.5.0
yarl==1.20.0
zipp==3.22.0
